Made showHints work for lots of windows.
The function showHints relied on a feature of UnderscoreJS that apparently
LoDash doesn't support. (The callback for groupBy assumed that the index of
the current element would be passed as its second argument.  Underscore does
this, but LoDash apparently does not.)

I'm not a great JS programmer. I ended up changing the partitioning to use
JavaScript's built-in Array.reduce and Object.values functions instead of
UnderscoreJS's (now LoDash's) _.groupBy and _.toArray.